refactor: file organization and shared patterns (SwiftUI Phase 2)
Extract StatusBarState to its own file, consolidate 16 inline NSCursor.pointingHand call sites into a shared PointingHandModifier, rename FileTreeHeaderContent/GitStatusHeaderContent to *View, and move tab-ordering logic from TabBarView to TabBarState for testability.
