+++ macos/Sources/Views/Modifiers/PointingHandModifier.swift
@@ -1,0 +1,42 @@
+import SwiftUI
+
+struct PointingHandModifier: ViewModifier {
+    var isEnabled: Bool = true
+    @State private var isHovered = false
+    @State private var didPushCursor = false
+
+    func body(content: Content) -> some View {
+        content
+            .onHover { hovering in
+                isHovered = hovering
+                syncCursor()
+            }
+            .onChange(of: isEnabled) { _, _ in
+                syncCursor()
+            }
+            .onDisappear {
+                if didPushCursor {
+                    NSCursor.pop()
+                    didPushCursor = false
+                }
+            }
+    }
+
+    private func syncCursor() {
+        let shouldPush = isHovered && isEnabled
+
+        if shouldPush && !didPushCursor {
+            NSCursor.pointingHand.push()
+            didPushCursor = true
+        } else if !shouldPush && didPushCursor {
+            NSCursor.pop()
+            didPushCursor = false
+        }
+    }
+}
+
+extension View {
+    func pointingHandCursor(isEnabled: Bool = true) -> some View {
+        modifier(PointingHandModifier(isEnabled: isEnabled))
+    }
+}
